+++ Dockerfile
@@ -7,4 +7,4 @@ RUN go mod download && go build
 # FROM scratch
 # COPY --from builder /build/hackaton-2020 /bin/hackaton-2020
 
-CMD ["/build/hackaton-2020"]
+CMD ["/build/server-sent-events-service"]
