+++ main.go
@@ -8,33 +8,30 @@ import (
 	"os"
 	"time"
 
+	"github.com/Hyperkid123/server-sent-events-service/src/enhancers"
+	"github.com/Hyperkid123/server-sent-events-service/src/kafkaconnector"
+	"github.com/Hyperkid123/server-sent-events-service/src/sse"
+	"github.com/Hyperkid123/server-sent-events-service/src/topics"
 	"github.com/gobuffalo/packr"
 	"github.com/joho/godotenv"
 	"gopkg.in/confluentinc/confluent-kafka-go.v1/kafka"
 )
 
-type Topics struct {
-	Topic     string   `json:"topic"`
-	Room      string   `json:"room"`
-	Event     string   `json:"event"`
-	Enhancers []string `json:"enhancers"`
-}
-
-func readTopics() map[string]Topics {
+func readTopics() map[string]topics.Topics {
 	box := packr.NewBox("./static")
 	file, foundError := box.FindString("topics.json")
 	if foundError != nil {
 		fmt.Println("Error while fetching file")
 		file = os.Getenv("CONFIG_JSON")
 	}
-	data := make([]Topics, 0)
+	data := make([]topics.Topics, 0)
 
 	err := json.Unmarshal([]byte(file), &data)
 	if err != nil {
 		fmt.Println(err)
 	}
 
-	dataMap := make(map[string]Topics)
+	dataMap := make(map[string]topics.Topics)
 	for i := 0; i < len(data); i++ {
 		dataMap[data[i].Topic] = data[i]
 	}
@@ -42,31 +39,31 @@ func readTopics() map[string]Topics {
 	return dataMap
 }
 
-func sendToListener(kafkaMessage *kafka.Message, topic Topics) {
+func sendToListener(kafkaMessage *kafka.Message, topic topics.Topics) {
 	if topic.Event == "" {
 		topic.Event = "notification"
 	}
 
 	enhancers := map[string](func(string, string) bool){
-		"inventory": InventoryEnhancer,
-		"approval":  ApprovalEnhancer,
+		"inventory": enhancers.InventoryEnhancer,
+		"approval":  enhancers.ApprovalEnhancer,
 	}
 	go func() {
-		for messageChannel, connectorInfo := range MessageChannels {
+		for messageChannel, connectorInfo := range sse.MessageChannels {
 			canSend := true
 			for i := 0; i < len(topic.Enhancers); i++ {
-				canSend = enhancers[topic.Enhancers[i]](string(kafkaMessage.Value), connectorInfo.accountNumber)
+				canSend = enhancers[topic.Enhancers[i]](string(kafkaMessage.Value), connectorInfo.AccountNumber)
 			}
 			if canSend {
-				msg := FormatSSE(topic.Event, string(kafkaMessage.Value))
+				msg := sse.FormatSSE(topic.Event, string(kafkaMessage.Value))
 				if topic.Room == "" {
 					fmt.Println("No room, broadcasting!")
 					messageChannel <- msg
-				} else if connectorInfo.room == topic.Room {
+				} else if connectorInfo.Room == topic.Room {
 					fmt.Println("Sending to specific room")
 					messageChannel <- msg
 				} else {
-					fmt.Println("Not sending", connectorInfo.room, topic.Room)
+					fmt.Println("Not sending", connectorInfo.Room, topic.Room)
 				}
 			}
 		}
@@ -89,9 +86,9 @@ func main() {
 		appName = "notifier"
 	}
 
-	go ConnectKafka(topicsConfig, sendToListener)
+	go kafkaconnector.ConnectKafka(topicsConfig, sendToListener)
 
-	http.HandleFunc(fmt.Sprintf("/api/%s/%s/connect", appName, apiVersion), ListenHandler)
+	http.HandleFunc(fmt.Sprintf("/api/%s/%s/connect", appName, apiVersion), sse.ListenHandler)
 	http.HandleFunc(fmt.Sprintf("/api/%s/%s/lubdub", appName, apiVersion), func(w http.ResponseWriter, r *http.Request) {
 		fmt.Fprintf(w, "lubdub")
 	})
