+++ src/enhancers/approval.go
@@ -1,4 +1,4 @@
-package main
+package enhancers
 
 import (
 	"fmt"
